requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
fake-useragent==1.5.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
pydantic==2.7.4
pydantic_core==2.18.4
pydotenv==0.0.7
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
python-2captcha